tests: Disable some portage sandbox features
To get rid of the Unable to unshare: EPERM (for FEATURES='ipc-sandbox network-sandbox pid-sandbox') messages.
This is required because the tests are run in a container and the container doesn't have the required permissions to make use of the underlying kernel features and we don't want to run in privileged mode.
See also https://bugs.gentoo.org/680456

Also unify the portage/make.conf FEATURES passed to the emerge and repoman tests, there's no real reason to have separate options for them