tests: Add tests to emerge all ebuilds
Split in two, one for emering all non-live (i.e. stable and testing) ebuilds and another to emerge all live ebuilds.
This will emerge all ebuilds in one go in a container meaning it will be faster than emerging all of them in a separate container but it's slightly less deterministic because the install from one of the earlier ebuilds might impact/interfere with the install of a later ebuild.
Because of this the test configs have been adapted as well to all write their configs to the same file to ensure that there are no USE flag blockers