ci: use pkgcheck instead of repoman
Repoman is dead and pkgcheck is its replacement for QA checks. This
approach also simplifies the checking of pull requests since pkgcheck
has builtin checking of changes in commit ranges. Thus, this commit not
just does a plain replacement, but also greatly simplifies the pkgcheck
worker script to just run the commit with the given options and process
the output once instead of for multiple repoman runs as was done
previously.