x11-libs/ntk: add patch to configure libdir in .pc files
The library search path in the .pc files was previously hard coded to
${prefix}/lib and thus was wrong for 17.1 amd64 profile Gentoo systems
where libdir = lib64.

This commit adds a patch to the build system and .pc.in templates to
use the libdir supplied at configure time instead of the hard coded
value. Note that this patch won't be sent upstream since the original
upstream is dead.

Closes: https://github.com/gentoo-audio/audio-overlay/issues/526
Signed-off-by: Adrian Schollmeyer <nex+b-g-o@nexadn.de>