ci: use official Gentoo binhost
Gentoo now provides official binhosts. We should use them as they can
reduce pipeline runtime (and thus the frequency of pipeline failures due
to timeouts). As the official binhost packages are distributed over the
Gentoo mirror infrastructure, these are highly available and thus also
solve the issue of the pipeline depending on a single working binhost to
function properly.

The binhost configuration is already supplied in the gentoo/stage3
Docker image, but the keys for the signatures must be fetched using
getuto.